Select schema to validate based on data version
Instead of simply just trying to validate against every schema version
in turn, try to get the data's version according to every schema version
in turn, and validate with only the one that's matching. If none match,
validate against the newest one.

This speeds up validation of data with older schemas, and simplifies
validation errors.

Fixes #1