v4: Rename "description" fields to "comment"
Replace "description" with "comment" as the field containing any extra
explanation the submitter wants to attach to the object, with the intent
to show it alongside the generated description in reports/dashboards.

The "description" name means it describes the object overall. However we
have other fields describing them in detail, and we'd rather use that to
generate our own description, consistently, regardless of the submitter.

Fixes #10