v4: Rename patch_mboxes to patchset_files
To prepare for addition of a patchset hash field, rename the
"patch_mboxes" field to "patchset_files", so they make better sense
together and appear related.

Relax the requirements towards the linked files to be acceptable by "git
apply", and not just "git am". This allows representing revisions with
arbitrary patches on top.