v4: Add log_excerpt fields
Add support for "log_excerpt" fields for "checkout" and "build" objects.
The field can contain an excerpt of the log of the checkout/build
attempt, that is most relevant to the outcome. E.g. patch application
failure messages, or particular errors which resulted in build failure.

The fields are capped at an arbitrary 16KB, just to start with
something. We can adjust that later.

This would help accommodate Intel's 0-Day with their log snippets, and
in general allow providing a summary of what happened with the
checkout/build, right in the notification message or the dashboard page.