Make version classes, not instances
Switch from expressing schema versions as instances of a single class,
to classes sharing a metaclass, and each version deriving from the
previous one.

This makes it easier to override both version and meta-version behavior
in the future, without having to define the logic for all versions in a
single class. It also makes it easier to reason about behavior changes
across versions.

Use this moment to make merging upgrade to the latest version across
merged data, and not the absolute latest version.