schema: Clarify uniqueness requirement for test IDs
We require that submitters provide unique IDs for each test they
are reporting but currently the documentation for the test schema
talks about test runs which could be interpreted as referring to
a test job as whole rather than a particular test.  Reword the
description to clarify things for submitters.

Signed-off-by: Mark Brown <broonie@kernel.org>