schema: Add MISS test status to v4.2
Accept a new string "MISS" as a test status in schema v4.2.
The new status is supposed to describe the situation where the test
harness planned to run the test, but failed to do that. E.g. due to
an infrastructure or provisioning failure.

Rewrite the description of the status field so the descriptions of the
various acceptable strings make better sense together, and the
description is easier to understand as a whole.