schema: Add (issue presence) transitions to v4.4
Add support for describing issue presence transitions.

The transitions point out the range of revisions along the change
history, where an issue has appeared or disappeared from the code. Note
that although there are no explicit revision objects in the I/O schema
they exist as aggregated objects in the ORM and the dashboards.

Transitions are generally expected to be substantiated by issue incidents
(which can be inferred from the revision and the issue linked from the
transition). However, that's not really necessary, as long as the
submitter makes it clear what happened in the transition and/or issue
description (or perhaps in the linked report), and makes sure it really
did.

Transitions can be "updated" - new versions of them submitted, when e.g.
they're located more precisely (the revision range was narrowed), or
for example they need to point to another issue.