schema: Support numerical test output in v4.4
Add support for specifying the numerical output produced by a test.

The meaning of the output is determined by the particular test. Should
only be considered when the test has actually executed. That is, with a
"FAIL", "ERROR", "PASS", or "DONE" status only. Normally "DONE" should
be used, when it's the main and not an auxiliary output.

The output "value" is expected to be floating-point, and the receiving
system must dedicate at least 64 bits for its storage.

Along the value it is possible to specify the "unit", as an arbitrary
string containing the unit symbol, or a (C-like) expression for a
compound unit, composed of base unit symbols.

If the "prefix" field is specified, the unit is considered to not
contain any prefix, the value is scaled as needed, and the appropriate
prefix (metric or binary) is added on display.