schema: Support test output values in v4.4
Add support for specifying the value measured by the test.

The meaning of the value is determined by the particular test. Should
only be considered when the test has actually executed. That is, with a
"FAIL", "ERROR", "PASS", or "DONE" status only. Normally "DONE" should
be used, when the value is the test's output, and not an auxilary value.

Each property inside the value corresponds to a data type. At least one
must be specified, but if more than one type property is set, each is
considered a different representation of the *same* value, and not a
different value. E.g. these can be specified at the same time:
"integer": 1, "number": 1, "boolean": true, "string": "true".

Specifying multiple types at once can be used to assist transitioning
test output to a different type, but in normal use only one of them
should be supplied.