schema: v5.0: Remove 'contacts' from checkouts
Remove the 'contacts' field from checkouts. It was added to support
extracting email addresses from patches so that notifications could be
sent to people concerned with them. Since then we've focused on
subscription-based system to minimize negative effect of sending false
positives/negatives to people who didn't ask for them.

So this field is not needed now. That is unless we start getting perfect
test results, which is not likely to ever happen.