schema: v5.0: Replace waived field with an incident
The "waived" field for tests can be generalized using "issues" and
"incidents", so remove it to simplify notifications and dashboards a
whole lot.

Whenever a test with `waived` set to true is inherited, add a
corresponding incident, with ID based on the test ID, linking it to a
global (constant-ID) issue.