schema: Support deduplication
Add a `dedup(data)` method to the abstract schema, which merges objects
with the same IDs right in a dataset, without the need to load it into a
database.

This allows our KCIDB-handling code in [Kwai][1] to build a dictionary
of objects and walk and process them without a database.

[1]: https://gitlab.com/cki-project/experimental/kwai