Support deduplication (#101)
* abstract: Wrap a long line

* actions: Run on latest Ubuntu

* actions: Drop KCIDB test triggering

Remove triggering KCIDB testing with kcidb-io version being tested, as
the former is not being developed or used anymore.

* schema: Support deduplication

Add a `dedup(data)` method to the abstract schema, which merges objects
with the same IDs right in a dataset, without the need to load it into a
database.

This allows our KCIDB-handling code in [Kwai][1] to build a dictionary
of objects and walk and process them without a database.

[1]: https://gitlab.com/cki-project/experimental/kwai